check remaining bytes before splitting
Bytes::split_to will panic if not enough bytes available. We do not want
panics so we do the checking first to make sure enough bytes available.
If not enough bytes available, it will send out a TruncatedMsg error.
The current iterator will catch the error and skip the remainder of the message.