removing output prefix for communities
previously, we have put prefixes like `lg:` or `ecv6` to indicating the
exact type of community it is. however, this is not a standard output
and users prefer to see just simple colon-separated number strings as
community values instead. this is particular the case for large
communities where other tools like bgpdump do not have prefixes, making
bgpkit-parser output non-comformant to popular tools.

this is potentially a breaking change for users who relies on elem
outputs instead of matching the actual community values to detect
specific community types.